+++ src/ui/assignments/src/pages/my-orders/AssignedOrders.tsx
@@ -1,55 +1,28 @@
-import { useEffect, useState, useRef } from "react";
-import {
-  Card,
-  CardHeader,
-  CardPreview,
-  Body1Stronger,
-  Divider,
-  Subtitle1,
-} from "@fluentui/react-components";
+import React, { useEffect, useState } from "react";
+import { Body1Stronger, Divider } from "@fluentui/react-components";
 
-import { useOrdersService } from "../../services/Orders.srv.ts";
+import { useOrdersService } from "../../services/Orders.service.ts";
 import {
   DONE_STATUS,
   FilteredOrder,
   FilteredSubItem,
-  isVisible,
   showOrder,
-  filterOrdersByText,
-  filterOrdersByType,
-  filterOrdersByProduct,
-  filterOrdersByDone,
 } from "../../services/Filters.service.ts";
-import { Loading } from "../../components/Loading.tsx";
-import { AppHeader } from "../../components/AppHeader.tsx";
-import { SubHeader, SubHeader2 } from "../../components/SubHeader.tsx";
 import {
   ConfirmDialog,
   ConfirmDialogProps,
 } from "../../components/ConfirmDialog.tsx";
-import { OrderComments, WithNote } from "../../components/OrderComments.tsx";
-import { Pagination } from "../../components/Pagination.tsx";
-import { Filters } from "../../components/filters/Filters.tsx";
-import { ItemsCount } from "../../components/ItemsCount.tsx";
 import { AssignedSubItems } from "./AssignedSubItems.tsx";
+import { BasePage } from "../../components/BasePage.tsx";
 
-const PAGE_SIZE = 25;
-
-export const AssignedOrders = () => {
-  // all assigned orders
-  const [myOrders, setMyOrders] = useState<(FilteredOrder & WithNote)[] | null>(
-    null
-  );
-
+export const AssignedOrders: React.FunctionComponent = () => {
+  const [myOrders, setMyOrders] = useState<FilteredOrder[] | null>(null); // assigned-orders list from the server
   const [statusesList, setStatusesList] = useState<string[] | undefined>(); // static list of status from the server
-  const [itemOffset, setItemOffset] = useState(0); // paging offset, display items from this index
 
   const [confirmOpen, setConfirmOpen] = useState<boolean | undefined>(false);
   const [confirmProps, setConfirmProps] = useState<
     Omit<ConfirmDialogProps, "openState"> | undefined
   >();
-
-  const pageRef = useRef<HTMLDivElement>(null);
 
   const ordersService = useOrdersService();
 
@@ -73,29 +46,6 @@ export const AssignedOrders = () => {
         .then((_statusesList) => setStatusesList(_statusesList));
     }
   }, []);
-
-  const _applyFilter = (filteredOrders: FilteredOrder[] | null) => {
-    if (filteredOrders !== null) {
-      setMyOrders(filteredOrders);
-      setItemOffset(0); // reset paging
-    }
-  };
-
-  const handleTilterByText = (searchText: string) => {
-    _applyFilter(filterOrdersByText(myOrders, searchText));
-  };
-
-  const handleFilterByType = (optionValue?: string) => {
-    _applyFilter(filterOrdersByType(myOrders, optionValue));
-  };
-
-  const handleFilterByProduct = (productsNames: string[]) => {
-    _applyFilter(filterOrdersByProduct(myOrders, productsNames));
-  };
-
-  const handleFilterByDone = (checked: boolean) => {
-    _applyFilter(filterOrdersByDone(myOrders, checked));
-  };
 
   const handleSubItemStatusChange = (
     orderId: number,
@@ -204,146 +154,48 @@ export const AssignedOrders = () => {
     setConfirmOpen(true);
   };
 
-  const toggleOpenNote = (orderId: number) => {
-    if (!myOrders) {
-      console.error("MyOrders::toggleOpenNote: orders empty");
-      return;
-    }
-
-    const orderIndex = myOrders.findIndex((order) => order.id === orderId);
-    if (orderIndex === -1) {
-      console.error("MyOrders::toggleOpenNote: order not found");
-      return;
-    }
-
-    myOrders[orderIndex].noteOpen = !myOrders[orderIndex].noteOpen;
-
-    myOrders.splice(orderIndex, 1, myOrders[orderIndex]);
-    setMyOrders([...myOrders]);
-  };
-
-  const handlePageClick = (pageIndex: number) => {
-    console.debug("MyOrders::handlePageClick", pageIndex);
-
-    const newOffset = (pageIndex * PAGE_SIZE) % filteredOrders.length;
-    setItemOffset(newOffset);
-
-    // scroll to top
-    pageRef.current?.scrollTo({ top: 0, behavior: "smooth" });
-  };
-
-  /** Orders match filter */
-  const filteredOrders = myOrders?.filter(isVisible) ?? [];
-
-  /** Orders to render on screen (based on paging) */
-  const visibleOrders = filteredOrders.slice(
-    itemOffset,
-    itemOffset + PAGE_SIZE
-  );
-
-  const pageNumber = Math.floor(itemOffset / PAGE_SIZE) + 1;
-  const pageCount = Math.ceil(filteredOrders.length / PAGE_SIZE);
-
   return (
     <>
-      {/* HEADER */}
-      <AppHeader />
-      {/* CONTENT */}
-      <div className="app-page" ref={pageRef}>
-        {/* SUB-HEADER */}
-        <SubHeader>הזמנות</SubHeader>
-        {!myOrders ? (
-          // SPINNER: LOADING...
-          <Loading />
-        ) : (
-          <div style={{ display: "flex", flexDirection: "column", gap: 12 }}>
-            {/* FILTERS */}
-            <Filters
-              onTextFilter={handleTilterByText}
-              onTypeFilter={handleFilterByType}
-              onProductFilter={handleFilterByProduct}
-              onDoneFilter={handleFilterByDone}
+      <BasePage
+        defaultOrders={myOrders}
+        title="הזמנות"
+        filters={{
+          text: true,
+          type: true,
+          product: true,
+          done: true,
+        }}
+        itemRender={{
+          header: (order) => (
+            <div style={{ width: "100%" }}>
+              <div>איזור: {order.region ?? "ללא איזור מוגדר"}</div>
+              <div>
+                איש קשר: {order.name}, טלפון:{" "}
+                <Body1Stronger>
+                  <a
+                    href={`tel:${
+                      (order.phone ?? "").startsWith("972") ? "+" : ""
+                    }${order.phone}`}
+                  >
+                    {order.phone}
+                  </a>
+                </Body1Stronger>
+              </div>
+              <Divider appearance="strong" />
+            </div>
+          ),
+          content: (order) => (
+            <AssignedSubItems
+              items={order.subItems}
+              statusesList={statusesList ?? []}
+              onStatusChange={(subItem, status, onCancel) =>
+                confirmSubItemStatusChange(order.id, subItem, status, onCancel)
+              }
+              onDelete={(item) => confirmSubItemRemove(order.id, item)}
             />
-            {myOrders.length === 0 ? (
-              // EMPTY STATE: NO DATA
-              <SubHeader2>אין הזמנות</SubHeader2>
-            ) : (
-              <>
-                {/* ITEMS COUNT */}
-                <ItemsCount
-                  itemName="הזמנות"
-                  itemsCount={myOrders.length}
-                  filteredItemsCount={filteredOrders.length}
-                  pageNumber={pageNumber}
-                  pageCount={pageCount}
-                />
-                {/* ITEMS LIST */}
-                {visibleOrders.filter(isVisible).map((order, index) => (
-                  <Card key={index}>
-                    <CardHeader
-                      header={
-                        <Subtitle1>{order.unit ?? "(ללא כותרת)"}</Subtitle1>
-                      }
-                      description={
-                        <div style={{ width: "100%" }}>
-                          <div>איזור: {order.region ?? "ללא איזור מוגדר"}</div>
-                          <div>
-                            איש קשר: {order.name}, טלפון:{" "}
-                            <Body1Stronger>
-                              <a
-                                href={`tel:${
-                                  (order.phone ?? "").startsWith("972")
-                                    ? "+"
-                                    : ""
-                                }${order.phone}`}
-                              >
-                                {order.phone}
-                              </a>
-                            </Body1Stronger>
-                          </div>
-                          <Divider appearance="strong" />
-                        </div>
-                      }
-                    />
-                    <CardPreview>
-                      <AssignedSubItems
-                        items={order.subItems}
-                        statusesList={statusesList ?? []}
-                        onStatusChange={(subItem, status, onCancel) =>
-                          confirmSubItemStatusChange(
-                            order.id,
-                            subItem,
-                            status,
-                            onCancel
-                          )
-                        }
-                        onDelete={(item) =>
-                          confirmSubItemRemove(order.id, item)
-                        }
-                      />
-                      <OrderComments
-                        order={order}
-                        onToggleNote={toggleOpenNote}
-                      />
-                    </CardPreview>
-                  </Card>
-                ))}
-                {/* PAGINATION */}
-                {filteredOrders.length > PAGE_SIZE && (
-                  <Pagination
-                    pageCount={pageCount}
-                    onPageClick={handlePageClick}
-                  />
-                )}
-                {/* EMPTY STATE: NO FILTER RESULTS */}
-                {filteredOrders.length === 0 && (
-                  <SubHeader2>אין הזמנות תואמת את הסינון</SubHeader2>
-                )}
-              </>
-            )}
-          </div>
-        )}
-      </div>
+          ),
+        }}
+      />
       {/* POPUP */}
       {confirmProps && (
         <ConfirmDialog
